refactor(worldmap): extract snapshot features as custom hooks
